Add consumer money and income

+++ src/slider/shaders/compute/producer.wgsl
@@ -4,8 +4,8 @@ fn main(@builtin(global_invocation_id) GlobalInvocationID : vec3<u32>) {
     if(index >= stats.num_producers) {
       return;
     }
-    let max_inventory = i32(producers[index].max_inventory);
-    var production_rate = i32(producers[index].production_rate);
+    let max_inventory = producers[index].max_inventory;
+    var production_rate = producers[index].production_rate;
 
     let old_a_inventory = atomicAdd(&producers[index].available_inventory, production_rate);
     let old_inventory = atomicAdd(&producers[index].inventory, production_rate);
